fix: add dump_tensors argument when running compiler for accuracy tests; cleanup debug logging; update spmm meta data type

+++ src/ArchitectureParams.h
@@ -125,7 +125,7 @@ using F9 = StdFloat<3, 5>;
 #define VU_INPUT_TYPES \
   VECTOR_DATATYPE, SCALE_DATATYPE, DataTypes::e4m3, DataTypes::int1
 #define OUTPUT_DATATYPES INPUT_DATATYPE, VU_INPUT_TYPES
-#define SPMM_META_DATATYPE DataTypes::int16
+#define SPMM_META_DATATYPE DataTypes::int32
 
 #define SA_INPUT_TYPE DataTypes::int6
 #define SA_WEIGHT_TYPE DataTypes::int6
